+++ packages/contracts/src/systems/S_SpendResourcesSystem.sol
@@ -10,7 +10,7 @@ import { EBuilding } from "src/Types.sol";
 import { BuildingKey } from "src/Keys.sol";
 
 contract S_SpendResourcesSystem is PrimodiumSystem {
-  function spendRequiredResources(bytes32 buildingEntity, uint32 level) public {
+  function spendRequiredResources(bytes32 buildingEntity, uint256 level) public {
     LibResource.spendRequiredResources(buildingEntity, level);
   }
 }
